requirements.txt
canonicaljson==2.0.0
python-dotenv==1.0.1
web3==6.20.1
requests==2.32.3
pytz==2024.2
python-dateutil==2.9.0.post0
